# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Open Api Document Server(Swagger)",
        "title": "open API(Swagger)",
        "contact": {
            "email": "harimkim@hyundai-autoever.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:18080",
    "basePath": "/",
    "paths": {
        "/vehiclemodel": {
            "get": {
                "description": "VehicleModelList를 조회한다",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a VehicleModelList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "모델명",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "제조사",
                        "name": "brand",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "제조사[]",
                        "name": "brands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "연료 유형",
                        "name": "fuelType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "연료 유형[]",
                        "name": "fuelTypes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "차량 등급",
                        "name": "grade",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "차량 등급[]",
                        "name": "grades",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vehiclemodel"
                        }
                    }
                }
            },
            "post": {
                "description": "VehicleModel을 생성한다.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a VehicleModelInsert",
                "parameters": [
                    {
                        "description": "모델명",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "제조사",
                        "name": "brand",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "대표 모델 여부",
                        "name": "standard",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "대표 모델 ID",
                        "name": "standardModelID",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "좌석수",
                        "name": "seatingCapacity",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "연료 유형",
                        "name": "fuelType",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "연비",
                        "name": "fuelEfficiency",
                        "in": "body",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "description": "주행 거리",
                        "name": "displacement",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "차량 등급",
                        "name": "grade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "웜업 시간",
                        "name": "warmUpTime",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vehiclemodel"
                        }
                    }
                }
            }
        },
        "/vehiclemodel/{vehiclemodelID}": {
            "get": {
                "description": "VehicleModel를 조회한다",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a  VehicleModelGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "모델 ID",
                        "name": "vehicleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vehiclemodel"
                        }
                    }
                }
            },
            "delete": {
                "description": "VehicleModel을 삭제한다.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a  VehicleModelDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "모델 ID",
                        "name": "vehicleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vehiclemodel"
                        }
                    }
                }
            },
            "patch": {
                "description": "VehicleModel을 수정한다.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a VehicleModelUpdate",
                "parameters": [
                    {
                        "description": "모델명",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "제조사",
                        "name": "brand",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "대표 모델 여부",
                        "name": "standard",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "대표 모델 ID",
                        "name": "standardModelID",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "좌석수",
                        "name": "seatingCapacity",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "연료 유형",
                        "name": "fuelType",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "연비",
                        "name": "fuelEfficiency",
                        "in": "body",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "description": "주행 거리",
                        "name": "displacement",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "차량 등급",
                        "name": "grade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "웜업 시간",
                        "name": "warmUpTime",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "type": "string",
                        "description": "모델 ID",
                        "name": "vehicleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vehiclemodel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Image": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "description": "1MB",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Vehiclemodel": {
            "type": "object",
            "properties": {
                "brand": {
                    "description": "제조사 이름",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "creatorName": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "displacement": {
                    "description": "배기량 cc",
                    "type": "integer"
                },
                "fuelEfficiency": {
                    "description": "연비",
                    "type": "number"
                },
                "fuelTankCapacity": {
                    "description": "연료탱크용량 liter",
                    "type": "integer"
                },
                "fuelType": {
                    "description": "연료 유형 : 전기",
                    "type": "string"
                },
                "grade": {
                    "description": "차량 등록 경형/소형/준중형/중형/대형/SUV",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "$ref": "#/definitions/model.Image"
                },
                "imageURL": {
                    "type": "string"
                },
                "name": {
                    "description": "모델명",
                    "type": "string"
                },
                "seatingCapacity": {
                    "description": "탑승인원",
                    "type": "integer"
                },
                "standard": {
                    "description": "표준 모델 여부",
                    "type": "boolean"
                },
                "standardModelID": {
                    "description": "표준 모델ID",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "warmUpTime": {
                    "description": "차량 이용 예비시간",
                    "type": "integer"
                }
            }
        }
    }
}